Перенес рендеринг доски и карточек в controllers/board.js
